+++ utils/types/types.ts
@@ -12,12 +12,12 @@ export interface KeyringAccount {
 export type LastEvent = {
   meta: string
   timestamp: string
-  nftId?: string
-  name?: string
-  issuer?: string
-  currentOwner?: string
-  metadata?: string
-  image?: string
+  nftId: string
+  name: string
+  issuer: string
+  currentOwner: string
+  metadata: string
+  image: string
   animationUrl: string
   collection?: string
 }
